requirements.txt

python-dotenv==1.0.0
twilio==8.5.0
